move comments to browser (#70319, #69103)
