+++ src/vs/workbench/api/electron-browser/extensionHost.contribution.ts
@@ -50,7 +50,7 @@ import '../browser/mainThreadTreeViews';
 import '../browser/mainThreadUrls';
 import '../browser/mainThreadWindow';
 import '../browser/mainThreadWorkspace';
-import './mainThreadComments';
+import '../browser/mainThreadComments';
 import './mainThreadTask';
 import './mainThreadWebview';
 import 'vs/workbench/api/node/apiCommands';
