+++ src/vs/workbench/contrib/comments/browser/commentGlyphWidget.ts
