+++ src/vs/workbench/contrib/comments/browser/comments.contribution.ts
@@ -6,8 +6,8 @@
 import * as nls from 'vs/nls';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { Registry } from 'vs/platform/registry/common/platform';
-import 'vs/workbench/contrib/comments/electron-browser/commentsEditorContribution';
-import { ICommentService, CommentService } from 'vs/workbench/contrib/comments/electron-browser/commentService';
+import 'vs/workbench/contrib/comments/browser/commentsEditorContribution';
+import { ICommentService, CommentService } from 'vs/workbench/contrib/comments/browser/commentService';
 import { IConfigurationRegistry, Extensions as ConfigurationExtensions } from 'vs/platform/configuration/common/configurationRegistry';
 
 export interface ICommentsConfiguration {
