+++ src/vs/workbench/contrib/comments/browser/media/panel.css
