+++ src/vs/workbench/contrib/comments/browser/simpleCommentEditor.ts
