+++ src/vs/workbench/workbench.main.ts
@@ -228,7 +228,7 @@ import 'vs/workbench/contrib/debug/browser/debugViewlet';
 import 'vs/workbench/contrib/markers/browser/markers.contribution';
 
 // Comments
-import 'vs/workbench/contrib/comments/electron-browser/comments.contribution';
+import 'vs/workbench/contrib/comments/browser/comments.contribution';
 
 // URL Support
 import 'vs/workbench/contrib/url/common/url.contribution';
